Bookings/requirements.txt
asgiref==3.7.2
autoflake==2.2.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
Deprecated==1.2.14
Django==4.2.5
django-oauth-toolkit==2.3.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
idna==3.4
isort==5.12.0
jwcrypto==1.5.0
oauthlib==3.2.2
Pillow==10.0.1
psycopg2-binary==2.9.7
pycparser==2.21
pyflakes==3.1.0
PyJWT==1.7.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.5
wrapt==1.15.0